dags/footprint_network/requirements.txt
requests==2.31.0
pandas==2.0.3
boto3==1.34.55
duckdb==0.9.2
pyarrow==14.0.1
python-dotenv==1.0.0
